refactor to create separate app and main files
