+++ src/app.tsx
@@ -1,157 +1,21 @@
-import { useState, useEffect, useRef } from "react";
-import { View, StyleSheet } from "react-native";
-import { LatLng } from "react-native-maps";
 import { registerRootComponent } from "expo";
-import * as Location from "expo-location";
-import * as Cellular from "expo-cellular";
-import { LocationObject } from "expo-location";
 
-import AppHeader from "./components/app-header";
-import MapViewContainer from "./components/map-view-container";
-import RouteButtonContainer from "./components/route-button-container";
-import InfoContainer from "./components/info-container";
-import NavigatorTab from "./components/navigator-tab";
-import UserIdStorage from "./utils/user-id-storage";
+import Main from "./components/main";
+import IdStorage from "./utils/id-storage";
 import UserIdStorageContext from "./contexts/user-id-context";
+import RouteIdStorageContext from "./contexts/route-id-context";
 
-const userIdStorage = new UserIdStorage();
+const userIdStorage = new IdStorage("user-id");
+const routeIdStorage = new IdStorage("route-id");
 
-function App() {
-	const [, setErrorMsg] = useState<string | null>(null);
-	const [curLocation, setCurLocation] = useState<LocationObject | null>(null);
-	const [mobileNetCode, setMobileNetCode] = useState<string | null>(null);
-	const [routeCoordinates, setRouteCoordinates] = useState<Array<LatLng>>([]);
-	const [showRoute, setShowRoute] = useState<boolean>(true);
-	const [trackingInterval] = useState<number>(1000);
-	const [isTracking, setIsTracking] = useState<boolean>(false);
-	const [tripId, setTripId] = useState<string | null>(null);
-	const trackingInfoRef = useRef<() => void>();
-
-	/**
-	 * Requests permissions to use device location.
-	 * Gets device location and sets current location state if permissions is allowed.
-	 * Otherwise sets error message state which is not currently used anywhere.
-	 * Also cellular network operators MNC (Mobile Network Code) and sets network code state
-	 * using the interval according to the network code interval state.
-	 * MNC code is a null if SIM card is not at the device or there is no cellular service available.
-	 */
-	useEffect(() => {
-		(async () => {
-			const { status } = await Location.requestForegroundPermissionsAsync();
-			if (status !== "granted") {
-				setErrorMsg("Permission to access location was denied");
-				return;
-			}
-			const location = await Location.getCurrentPositionAsync({});
-			setCurLocation(location);
-			const networkCode = await Cellular.getMobileNetworkCodeAsync();
-			setMobileNetCode(networkCode);
-		})();
-	}, []);
-
-	/**
-	 * Gets devices last known location and MNC code, updates corresponding states and appends
-	 * the coordinate points (latitude/longitude) into route coordinates state using the
-	 * `AddNewRouteCoordinate()`function.
-	 */
-	const updateTrackingInfo: () => void = async () => {
-		const location = await Location.getLastKnownPositionAsync({});
-		setCurLocation(location);
-		addNewRouteCoordinate(location);
-		const networkCode = await Cellular.getMobileNetworkCodeAsync();
-		setMobileNetCode(networkCode);
-	};
-
-	/**
-	 * Updates tracking info using the interval of tracking interval state if
-	 * is tracking state has been set to true. Otherwise do not update.
-	 */
-	useEffect(() => {
-		trackingInfoRef.current = updateTrackingInfo;
-	}, [updateTrackingInfo]);
-
-	useEffect(() => {
-		function tick() {
-			trackingInfoRef.current();
-		}
-		if (isTracking) {
-			const id = setInterval(tick, trackingInterval);
-			return () => clearInterval(id);
-		}
-	}, [isTracking]);
-
-	/**
-	 * Creates coordinate object from the location object and appends coordinate into
-	 * route coordinate state which contains list of previously stored coordinate objects.
-	 * @param location object which contains latitude and longitude values.
-	 */
-	const addNewRouteCoordinate = (location: LocationObject | null) => {
-		if (location !== null) {
-			const coordinate = {
-				latitude: location.coords.latitude,
-				longitude: location.coords.longitude,
-			};
-			setRouteCoordinates(routeCoordinates.concat(coordinate));
-		}
-	};
-
-	/**
-	 * Changes tracking state, initializes route coordinate state with empty list
-	 * and sets trip id to string when tracking has started and to null when tracking
-	 * has ended.
-	 */
-	const changeTracking = () => {
-		// Temporary solution. Waiting for backend connection and userID request.
-		const uid = "non-unique-trip-id-1234";
-		setRouteCoordinates([]);
-		setIsTracking(!isTracking);
-		setTripId(tripId ? null : uid);
-		console.log(isTracking ? "tracking started" : "tracking ended");
-	};
-
-	/**
-	 * Changes show route state to opposite boolean ie. from true to false
-	 * and vice versa.
-	 */
-	const changeShowRoute = () => {
-		console.log(showRoute);
-		setShowRoute(!showRoute);
-	};
-
+const App = () => {
 	return (
 		<UserIdStorageContext.Provider value={userIdStorage}>
-			<View style={styles.container}>
-				<AppHeader name={"Berry picker tracker"} />
-				<MapViewContainer
-					showRoute={showRoute}
-					routeCoordinates={routeCoordinates}
-				/>
-				<RouteButtonContainer
-					changeTracking={changeTracking}
-					changeShowRoute={changeShowRoute}
-					showRoute={showRoute}
-					isTracking={isTracking}
-					tripId={tripId}
-				/>
-				<InfoContainer
-					curLocation={curLocation}
-					mobileNetCode={mobileNetCode}
-					routeCoordinates={routeCoordinates}
-				/>
-				<NavigatorTab />
-			</View>
+			<RouteIdStorageContext.Provider value={routeIdStorage}>
+				<Main />
+			</RouteIdStorageContext.Provider>
 		</UserIdStorageContext.Provider>
 	);
-}
-
-const styles = StyleSheet.create({
-	container: {
-		flex: 1,
-		backgroundColor: "#fff",
-		alignItems: "center",
-		justifyContent: "flex-start",
-		flexDirection: "column",
-	},
-});
+};
 
 export default registerRootComponent(App);
